19: Moving SocialLink(s) components to folder structure
