+++ src/components/SocialLinks/SocialLink/style.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const SocialLinkListItem = styled.li`
+  margin-bottom: 0;
+`;
+
+export const SocialLinkIcon = styled.i`
+  font-size: 26px;
+  line-height: 26px;
+  color: #5a6e87;
+  transition: color 0.5s;
+
+  &:hover {
+    color: #4d7ab3;
+  }
+`;
