+++ src/components/SocialLinks/style.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const SocialLinksList = styled.ul`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: auto;
+  max-width: 300px;
+  list-style: none;
+`;
